pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "indication-scout"
version = "0.1.0"
description = "A machine learning toolkit for drug repurposing and indication discovery"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Dhruv Gupta" }
]
keywords = [
    "drug-repurposing",
    "machine-learning",
    "biomedical",
    "indication-discovery",
    "evidence-ranking"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "click>=8.1.0",
    "black>=23.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.9.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3.0",
    "sentence-transformers>=3.0.0",
    "deepeval>=3.4.3",
    "wandb>=0.18.0",
    "langchain-core>=1.2.23",
    "langchain>=1.2.13",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
all = [
    "indication-scout[dev,notebooks]",
]

[project.scripts]
scout = "indication_scout.cli.cli:main"

[project.urls]
Homepage = "https://github.com/dgupta/indication-scout"
Repository = "https://github.com/dgupta/indication-scout"
Issues = "https://github.com/dgupta/indication-scout/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
indication_scout = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["indication_scout"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
exclude = [
    "\\.venv",
    "build",
    "dist",
]

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "pandas.*",
    "numpy.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/indication_scout"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
